% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_aod_wald_test.R
\name{ard_aod_wald_test}
\alias{ard_aod_wald_test}
\title{ARD Wald Test}
\usage{
ard_aod_wald_test(x, ...)
}
\arguments{
\item{x}{regression model object}

\item{...}{arguments passed to \code{aod::wald.test(...)}}
}
\value{
data frame
}
\description{
Function takes a regression model object and calculates Wald
statistical test using \code{\link[aod:wald.test]{aod::wald.test()}}.
}
\examples{
\dontshow{if (cards::is_pkg_installed(c("aod"), reference_pkg = "cardx")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
lm(AGE ~ ARM, data = cards::ADSL) |>
  ard_aod_wald_test()
\dontshow{\}) # examplesIf}
}
