% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_aov.R
\name{ard_aov}
\alias{ard_aov}
\title{ARD ANOVA}
\usage{
ard_aov(x, data, ...)
}
\arguments{
\item{x}{linear model formula}

\item{data}{(\code{data.frame})\cr
a data frame containing the variables in \code{x}}

\item{...}{arguments passed to \code{stats::aov(...)}}
}
\value{
ARD data frame
}
\description{
Analysis results data for Analysis of Variance.
Calculated with \code{stats::aov()}
}
\examples{
ard_aov(AGE ~ ARM, data = cards::ADSL)
}
