% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_categorical_max.R
\name{ard_categorical_max}
\alias{ard_categorical_max}
\title{ARD to Calculate Categorical Occurrence Rates by Maximum Level Per Unique ID}
\usage{
ard_categorical_max(
  data,
  variables,
  id,
  by = dplyr::group_vars(data),
  statistic = everything() ~ c("n", "p", "N"),
  denominator = NULL,
  fmt_fn = NULL,
  stat_label = everything() ~ cards::default_stat_labels(),
  quiet = TRUE,
  ...
)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame}

\item{variables}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
The categorical variables for which occurrence rates per unique ID (by maximum level) will be calculated.}

\item{id}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Argument used to subset \code{data} to identify rows in \code{data} to calculate categorical variable level occurrence rates.}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
columns to tabulate by in the series of ARD function calls.
Any rows with \code{NA} or \code{NaN} values are removed from all calculations.}

\item{statistic}{(\code{\link[cards:syntax]{formula-list-selector}})\cr
a named list, a list of formulas,
or a single formula where the list element one or more of  \code{c("n", "N", "p")}
(or the RHS of a formula).}

\item{denominator}{(\code{data.frame}, \code{integer})\cr
Used to define the denominator and enhance the output.
The argument is optional. If not specified, \code{data} will be used as \code{denominator}.
\itemize{
\item the univariate tabulations of the \code{by} variables are calculated with \code{denominator} when a data frame is passed,
e.g. tabulation of the treatment assignment counts that may appear in the header of a table.
}}

\item{fmt_fn}{(\code{\link[cards:syntax]{formula-list-selector}})\cr
a named list, a list of formulas,
or a single formula where the list element is a named list of functions
(or the RHS of a formula),
e.g. \verb{list(mpg = list(mean = \\(x) round(x, digits = 2) |> as.character()))}.}

\item{stat_label}{(\code{\link[cards:syntax]{formula-list-selector}})\cr
a named list, a list of formulas, or a single formula where
the list element is either a named list or a list of formulas defining the
statistic labels, e.g. \code{everything() ~ list(n = "n", p = "pct")} or
\code{everything() ~ list(n ~ "n", p ~ "pct")}.}

\item{quiet}{(scalar \code{logical})\cr
Logical indicating whether to suppress additional messaging. Default is \code{FALSE}.}

\item{...}{Arguments passed to methods.}
}
\value{
an ARD data frame of class 'card'
}
\description{
Function calculates categorical variable level occurrences rates by maximum level per unique ID.
Each variable in \code{variables} is evaluated independently and then results for all variables are stacked.
Only the highest-ordered level will be counted for each unique ID.
Unordered, non-numeric variables will be converted to factor and the default level order used for ordering.
}
\examples{
# Occurrence Rates by Max Level (Highest Severity) --------------------------
ard_categorical_max(
  cards::ADAE,
  variables = c(AESER, AESEV),
  id = USUBJID,
  by = TRTA,
  denominator = cards::ADSL |> dplyr::rename(TRTA = ARM),
  quiet = FALSE
)
}
