% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_continuous_ci.R
\name{ard_continuous_ci}
\alias{ard_continuous_ci}
\alias{ard_continuous_ci.data.frame}
\title{ARD continuous CIs}
\usage{
ard_continuous_ci(data, ...)

\method{ard_continuous_ci}{data.frame}(
  data,
  variables,
  by = dplyr::group_vars(data),
  conf.level = 0.95,
  method = c("t.test", "wilcox.test"),
  ...
)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame}

\item{...}{arguments passed to \code{t.test()} or \code{wilcox.test()}}

\item{(`string`)\cr}{a string indicating the method to use for the confidence interval
calculation. Must be one of \code{"t.test"} or \code{"wilcox.test"}}
}
\value{
ARD data frame
}
\description{
One-sample confidence intervals for continuous variable means and medians.
}
\examples{
ard_continuous_ci(mtcars, variables = c(mpg, hp), method = "wilcox.test")
ard_continuous_ci(mtcars, variables = mpg, by = am, method = "t.test")
}
