% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_dichotomous.survey.design.R
\name{ard_dichotomous.survey.design}
\alias{ard_dichotomous.survey.design}
\title{ARD Dichotomous Survey Statistics}
\usage{
\method{ard_dichotomous}{survey.design}(data, ...)
}
\arguments{
\item{data}{(\code{survey.design})\cr
a design object often created with \code{\link[survey:svydesign]{survey::svydesign()}}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{value}{(named \code{list})\cr
named list of dichotomous values to tabulate.
Default is \code{cards::maximum_variable_value(data$variables)},
which returns the largest/last value after a sort.}
}
\value{
an ARD data frame of class 'card'
}
\description{
Compute Analysis Results Data (ARD) for dichotomous summary statistics.
}
\examples{
\dontshow{if (cardx:::is_pkg_installed("survey")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
survey::svydesign(ids = ~1, data = mtcars, weights = ~1) |>
  ard_dichotomous(by = vs, variables = c(cyl, am), value = list(cyl = 4))
\dontshow{\}) # examplesIf}
}
