% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_hedges_g.R
\name{ard_hedges_g}
\alias{ard_hedges_g}
\alias{ard_paired_hedges_g}
\title{ARD Hedge's G Test}
\usage{
ard_hedges_g(data, by, variable, ...)

ard_paired_hedges_g(data, by, variable, id, ...)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame. See below for details.}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name to compare by. Must be a categorical variable with exactly two levels.}

\item{variable}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name to be compared. Must be a continuous variable.}

\item{...}{arguments passed to \code{effectsize::hedges_g(...)}}

\item{id}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name of the subject or participant ID}
}
\value{
ARD data frame
}
\description{
Analysis results data for paired and non-paired Hedge's G Effect Size Test
using \code{\link[effectsize:cohens_d]{effectsize::hedges_g()}}.
}
\details{
For the \code{ard_hedges_g()} function, the data is expected to be one row per subject.
The data is passed as \code{effectsize::hedges_g(data[[variable]]~data[[by]], data, paired = FALSE, ...)}.

For the \code{ard_paired_hedges_g()} function, the data is expected to be one row
per subject per by level. Before the effect size is calculated, the data are
reshaped to a wide format to be one row per subject.
The data are then passed as
\verb{effectsize::hedges_g(x = data_wide[[<by level 1>]], y = data_wide[[<by level 2>]], paired = TRUE, ...)}.
}
\examples{
cards::ADSL |>
  dplyr::filter(ARM \%in\% c("Placebo", "Xanomeline High Dose")) |>
  ard_hedges_g(by = ARM, variable = AGE)

# constructing a paired data set,
# where patients receive both treatments
cards::ADSL[c("ARM", "AGE")] |>
  dplyr::filter(ARM \%in\% c("Placebo", "Xanomeline High Dose")) |>
  dplyr::mutate(.by = ARM, USUBJID = dplyr::row_number()) |>
  dplyr::arrange(USUBJID, ARM) |>
  dplyr::group_by(USUBJID) |>
  dplyr::filter(dplyr::n() > 1) |>
  ard_paired_hedges_g(by = ARM, variable = AGE, id = USUBJID)
}
