% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_incidence_rate.R
\name{ard_incidence_rate}
\alias{ard_incidence_rate}
\title{ARD Incidence Rate}
\usage{
ard_incidence_rate(
  data,
  variable,
  cnsr,
  id,
  conf.level = 0.95,
  conf.type = "normal",
  var_unit = "year",
  n_person_years = 100
)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame.}

\item{variable}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name of time-to-event variable.}

\item{cnsr}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name used to indicate censoring, i.e. whether an event has been observed (1) or not (0).}

\item{id}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name used to identify unique patients in \code{data}.}

\item{conf.level}{(\code{numeric})\cr
confidence level for the estimated incidence rate.}

\item{conf.type}{(\code{string})\cr
confidence interval type for the estimated incidence rate.

One of: \code{normal} (default), \code{normal-log}, \code{exact}, or \code{byar}.}

\item{var_unit}{(\code{string})\cr
unit of time of values in \code{x}.

One of: \code{day}, \code{week}, \code{month}, or \code{year} (default)}

\item{n_person_years}{(\code{numeric})\cr
number of person-years to estimate incidence rate for. Defaults to 100.}
}
\value{
an ARD data frame of class 'card'
}
\description{
Function takes a time variable (\code{variable}), censoring variable (\code{cnsr}), and patient ID variable (\code{id}) and
calculates the estimated incidence rate in person years.
}
\examples{
cards::ADTTE |>
  ard_incidence_rate(AVAL, CNSR, USUBJID)

cards::ADTTE |>
  ard_incidence_rate(AVAL, CNSR, USUBJID, var_unit = "month", n_person_years = 10)
}
