% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_kwtest.R
\name{ard_kwtest}
\alias{ard_kwtest}
\title{ARD Kruskal-Wallis Test}
\usage{
ard_kwtest(data, by, variable, ...)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame.}

\item{by, }{variable (\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column names to compare}

\item{...}{additional arguments}
}
\value{
ARD data frame
}
\description{
Analysis results data for Kruskal-Wallis Rank Sum Test.
Calculated with \code{kruskal.test(formula = data[[variable]] ~ data[[by]], data = data, ...)}
}
\examples{
cards::ADSL |>
  ard_kwtest(by = "ARM", variable = "AGEGR1")
}
