% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_regression.R
\name{ard_regression_basic}
\alias{ard_regression_basic}
\title{ard_regression_basic}
\usage{
ard_regression_basic(
  x,
  tidy_fun = broom.helpers::tidy_with_broom_or_parameters,
  ...
)
}
\arguments{
\item{x}{regression model object}

\item{...}{Arguments passed to \code{broom.helpers::tidy_plus_plus()}}

\item{tidy_fun(`function`)\cr}{a tidier. Default is \code{\link[broom.helpers:tidy_with_broom_or_parameters]{broom.helpers::tidy_with_broom_or_parameters}}}
}
\value{
data frame
}
\description{
A function that takes a regression model and provides basic statistics in an
ARD structure.
}
\examples{
lm(AGE ~ ARM, data = cards::ADSL) |>
  ard_regression_basic()

}
