% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_single_ci.R
\name{ard_single_ci}
\alias{ard_single_ci}
\title{ARD Univariate Confidence Intervals}
\usage{
ard_single_ci(
  data,
  variables,
  by = dplyr::group_vars(data),
  conf.level = 0.95,
  method = c("mean_with_Z", "mean_with_T")
)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame}

\item{variables}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
columns to include in summaries. Columns must be class \verb{<logical>}
or \verb{<numeric>} values coded as \code{c(0, 1)}.}

\item{conf.level}{(\code{numeric})\cr
a scalar in \verb{(0, 1)} indicating the confidence level.
Default is \code{0.95}}

\item{method}{(\code{string})\cr
string indicating the type of confidence interval to calculate.
Must be one of 'waldcc', 'wald', 'clopper-pearson', 'wilson', 'wilsoncc', 'strat_wilson', 'strat_wilsoncc', 'agresti-coull', 'jeffreys'.
See \code{?single_ci} for details.}
}
\value{
an ARD data frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}\cr
Calculate confidence intervals for univariate statistics.
}
\examples{
\dontshow{if (cards::is_pkg_installed("broom", reference_pkg = "cardx")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ard_single_ci(cards::ADSL, variables = c(AGE), method = "mean_with_Z")
ard_single_ci(cards::ADSL, variables = c(AGE), method = "mean_with_T")
\dontshow{\}) # examplesIf}
}
