% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_smd.R
\name{ard_smd}
\alias{ard_smd}
\title{ARD Standardized Mean Difference}
\usage{
ard_smd(data, by, variable, ...)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame.}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name to compare by}

\item{variable}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name to be compared.}

\item{...}{
  Arguments passed on to \code{\link[smd:smd]{smd::smd}}
  \describe{
    \item{\code{std.error}}{Logical indicator for computing standard errors using
\code{\link[smd]{compute_smd_var}}. Defaults to \code{FALSE}.}
    \item{\code{gref}}{an integer indicating which level of \code{g} to use as the reference
group. Defaults to \code{1}.}
  }}
}
\value{
ARD data frame
}
\description{
Standardized mean difference calculated via \code{\link[smd:smd]{smd::smd()}} with \code{na.rm = TRUE}.
}
\examples{
ard_smd(cards::ADSL, by = ARM, variable = AGE, std.error = TRUE)
ard_smd(cards::ADSL, by = ARM, variable = AGEGR1, std.error = TRUE)
}
