% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_survest.R
\name{ard_survest}
\alias{ard_survest}
\title{ARD Survival Estimates}
\usage{
ard_survest(
  x,
  by = NULL,
  times = NULL,
  probs = NULL,
  conf.level = 0.95,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{x}{(\code{\link[survival:survfit]{survival::survfit()}})\cr
a \code{\link[survival:survfit]{survival::survfit()}} object. See below for details.}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name to compare by}

\item{...}{arguments passed to \code{t.test(...)}}

\item{variable}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name to be compared}

\item{id}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name of the subject or participant ID}
}
\value{
ARD data frame
}
\description{
Analysis results data for survival quantiles and x-year survival estimates, extracted
from a \code{\link[survival:survfit]{survival::survfit()}} model.
}
\details{
For the \code{ard_survest()} function, the data is expected to be one row per subject.
The data is passed as \code{t.test(data[[variable]] ~ data[[by]], paired = FALSE, ...)}.

For the \code{ard_paired_ttest()} function, the data is expected to be one row
per subject per by level. Before the t-test is calculated, the data are
reshaped to a wide format to be one row per subject.
The data are then passed as
\verb{t.test(x = data_wide[[<by level 1>]], y = data_wide[[<by level 2>]], paired = TRUE, ...)}.
}
\examples{
\dontshow{if (broom.helpers::.assert_package("survival", pkg_search = "cardx", boolean = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(survival)

survfit(Surv(AVAL, CNSR) ~ TRTA, cards::ADTTE) |>
  ard_survest()
\dontshow{\}) # examplesIf}
}
