% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_survey_categorical_ci.R
\name{ard_survey_categorical_ci}
\alias{ard_survey_categorical_ci}
\title{ARD survey categorical CIs}
\usage{
ard_survey_categorical_ci(
  data,
  variables,
  by = NULL,
  method = c("logit", "likelihood", "asin", "beta", "mean", "xlogit"),
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{data}{(\code{survey.design})\cr
a design object often created with \code{\link[survey:svydesign]{survey::svydesign()}}.}

\item{variables}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
columns to include in summaries. Default is \code{everything()}.}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
results are calculated for \strong{all combinations} of the columns specified,
including unobserved combinations and unobserved factor levels.}

\item{method}{(\code{string})\cr
Method passed to \code{survey::svyciprop(method)}}

\item{conf.level}{(scalar \code{numeric})\cr
confidence level for confidence interval. Default is \code{0.95}.}

\item{...}{arguments passed to \code{survey::svyciprop()}}
}
\value{
ARD data frame
}
\description{
Confidence intervals for categorical variables calculated via
\code{\link[survey:svyciprop]{survey::svyciprop()}}.
}
\examples{
\dontshow{if (do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = "survey", reference_pkg = "cardx"))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(api, package = "survey")
dclus1 <- survey::svydesign(id = ~dnum, weights = ~pw, data = apiclus1, fpc = ~fpc)

ard_survey_categorical_ci(dclus1, variables = sch.wide)
ard_survey_categorical_ci(dclus1, variables = sch.wide, method = "xlogit")
\dontshow{\}) # examplesIf}
}
