% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_survfit.R
\name{ard_survfit}
\alias{ard_survfit}
\title{ARD Survival Estimates}
\usage{
ard_survfit(x, times = NULL, probs = NULL, reverse = FALSE)
}
\arguments{
\item{x}{(\code{\link[survival:survfit]{survival::survfit()}})\cr
a \code{\link[survival:survfit]{survival::survfit()}} object. See below for details.}

\item{times}{(\code{numeric})\cr
a vector of times for which to return survival probabilities.}

\item{probs}{(\code{numeric})\cr
a vector of probabilities with values in (0,1) specifying the survival quantiles to return.}

\item{reverse}{(\code{logical})\cr
Flip the probability reported, i.e. \code{1 - estimate}. Default is \code{FALSE}. Only applies when
\code{probs} is specified.}
}
\value{
an ARD data frame of class 'card'
}
\description{
Analysis results data for survival quantiles and x-year survival estimates, extracted
from a \code{\link[survival:survfit]{survival::survfit()}} model.
}
\details{
\itemize{
\item Only one of either the \code{times} or \code{probs} parameters can be specified.
\item Times should be provided using the same scale as the time variable used to fit the provided
survival fit model.
}
}
\examples{
\dontshow{if (broom.helpers::.assert_package("survival", pkg_search = "cardx", boolean = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(survival)

survfit(Surv(AVAL, CNSR) ~ TRTA, cards::ADTTE) |>
  ard_survfit(times = c(60, 180))

survfit(Surv(AVAL, CNSR) ~ TRTA, cards::ADTTE) |>
  ard_survfit(probs = c(0.25, 0.5, 0.75))
\dontshow{\}) # examplesIf}
}
