% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_svychisq.R
\name{ard_svychisq}
\alias{ard_svychisq}
\title{ARD Survey Chi-Square Test}
\usage{
ard_svychisq(data, by, variable, statistic = "F", ...)
}
\arguments{
\item{data}{(\code{survey.design})\cr
a survey design object often created with}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name to compare by}

\item{variable}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name to be compared}

\item{statistic}{(\code{character})\cr
statistic used to estimate chisq p-value.
Default is the Rao-Scott second-order correction ("F"). See \code{\link[survey:svychisq]{survey::svychisq}}
for available statistics options.}

\item{...}{arguments passed to \code{\link[survey:svychisq]{survey::svychisq()}}}
}
\value{
ARD data frame
}
\description{
Analysis results data for survey Chi-Square test using \code{\link[survey:svychisq]{survey::svychisq()}}.
}
\examples{
\dontshow{if (cards::is_pkg_installed(c("survey", "broom"), reference_pkg = "cardx")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(api, package = "survey")
dclus2 <- survey::svydesign(id=~dnum, weights=~pw, data=apiclus1, fpc=~fpc)

ard_svychisq(dclus2, variable = sch.wide, by = comp.imp, statistic = "F")
\dontshow{\}) # examplesIf}
}
