% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_svycontinuous.R
\name{ard_svycontinuous}
\alias{ard_svycontinuous}
\title{ARD Continuous Survey Statistics}
\usage{
ard_svycontinuous(
  data,
  variables,
  by = NULL,
  statistic = everything() ~ c("median", "p25", "p75"),
  fmt_fn = NULL,
  stat_label = NULL
)
}
\arguments{
\item{data}{(\code{survey.design})\cr
a design object often created with \code{\link[survey:svydesign]{survey::svydesign()}}.}

\item{variables}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
columns to include in summaries. Default is \code{everything()}.}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
results are calculated for \strong{all combinations} of the columns specified,
including unobserved combinations and unobserved factor levels.}

\item{statistic}{(\code{\link[=syntax]{formula-list-selector}})\cr
a named list, a list of formulas,
or a single formula where the list element is a character vector of
statistic names to include. See below for options.}

\item{fmt_fn}{(\code{\link[=syntax]{formula-list-selector}})\cr
a named list, a list of formulas,
or a single formula where the list element is a named list of functions
(or the RHS of a formula),
e.g. \verb{list(mpg = list(mean = \\(x) round(x, digits = 2) |> as.character))}.}

\item{stat_label}{(\code{\link[=syntax]{formula-list-selector}})\cr
a named list, a list of formulas, or a single formula where
the list element is either a named list or a list of formulas defining the
statistic labels, e.g. \code{everything() ~ list(mean = "Mean", sd = "SD")} or
\code{everything() ~ list(mean ~ "Mean", sd ~ "SD")}.}
}
\value{
an ARD data frame of class 'card'
}
\description{
Returns an ARD of weighted statistics using the \code{{survey}} package.
}
\section{statistic argument}{


The following statistics are available:
"mean", "median", "min", "max", "sum", "var", "sd", "mean.std.error", "deff", "p##",
where 'p##' is are the percentiles and \verb{##} is an integer between 0 and 100.
}

\examples{
\dontshow{if (cards::is_pkg_installed("survey", reference_pkg = "cardx")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(api, package = "survey")
dclus1 <- survey::svydesign(id = ~dnum, weights = ~pw, data = apiclus1, fpc = ~fpc)

ard_svycontinuous(
  data = dclus1,
  variables = api00,
  by = stype
)
\dontshow{\}) # examplesIf}
}
