% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_svyttest.R
\name{ard_svyttest}
\alias{ard_svyttest}
\title{ARD Survey t-test}
\usage{
ard_svyttest(data, by, variable, conf.level = 0.95, ...)
}
\arguments{
\item{data}{(\code{survey.design})\cr
a survey design object often created with}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name to compare by}

\item{variable}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name to be compared}

\item{conf.level}{(\code{double})\cr
confidence level of the returned confidence interval. Must be between \code{c(0, 1)}.
Default is \code{0.95}}

\item{...}{arguments passed to \code{\link[survey:svyttest]{survey::svyttest()}}}
}
\value{
ARD data frame
}
\description{
Analysis results data for survey t-test using \code{\link[survey:svyttest]{survey::svyttest()}}.
}
\examples{
\dontshow{if (cards::is_pkg_installed(c("survey", "broom"), reference_pkg = "cardx")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(api, package = "survey")
dclus2 <- survey::svydesign(id = ~dnum + snum, fpc = ~fpc1 + fpc2, data = apiclus2)

ard_svyttest(dclus2, variable = enroll, by = comp.imp, conf.level = 0.9)
\dontshow{\}) # examplesIf}
}
