% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_vif.R
\name{ard_vif}
\alias{ard_vif}
\title{Regression VIF ARD}
\usage{
ard_vif(x)
}
\arguments{
\item{x}{regression model object
See car::vif() for details}

\item{...}{arguments passed to \code{car::vif(...)}}
}
\value{
data frame
}
\description{
Function takes a regression model object and converts it to a ARD
structure using the \code{broom.helpers} package.
}
\examples{
lm(AGE ~ ARM + SEX, data = cards::ADSL) |>
  ard_vif()
}
