% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{ard_continuous.survey.design}
\alias{ard_categorical.survey.design}
\alias{ard_dichotomous.survey.design}
\alias{ard_categorical_max}
\alias{ard_emmeans_mean_difference}
\title{Deprecated functions}
\usage{
\method{ard_continuous}{survey.design}(data, ...)

\method{ard_categorical}{survey.design}(data, ...)

\method{ard_dichotomous}{survey.design}(data, ...)

ard_categorical_max(...)

ard_emmeans_mean_difference(...)
}
\value{
See each individual function help page.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}\cr
Some functions have been deprecated and are no longer being actively
supported.
}
\examples{
NULL
}
\keyword{internal}
