% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_hedgesg.R
\name{.format_hedgesg_results}
\alias{.format_hedgesg_results}
\title{Convert Hedge's G Test to ARD}
\usage{
.format_hedgesg_results(by, variable, lst_tidy, paired, ...)
}
\arguments{
\item{by}{(\code{string})\cr by column name}

\item{variable}{(\code{string})\cr variable column name}

\item{lst_tidy}{(named \code{list})\cr
list of tidied results constructed with \code{\link[cards:eval_capture_conditions]{eval_capture_conditions()}},
e.g. \code{eval_capture_conditions(t.test(mtcars$mpg ~ mtcars$am) |> broom::tidy())}.}

\item{paired}{If \code{TRUE}, the values of \code{x} and \code{y} are considered as paired.
This produces an effect size that is equivalent to the one-sample effect
size on \code{x - y}.}

\item{...}{passed to \code{hedges_g(...)}}
}
\value{
ARD data frame
}
\description{
Convert Hedge's G Test to ARD
}
\examples{
cardx:::.format_hedgesg_results(
  by = "ARM",
  variable = "AGE",
  paired = FALSE,
  lst_tidy =
    cards::eval_capture_conditions(
      effectsize::hedges_g(data[[variable]] ~ data[[by]], paired = FALSE) |>
        parameters::standardize_names(style = "broom")
    )
)
}
\keyword{internal}
