% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_stats_prop_test.R
\name{.format_proptest_results}
\alias{.format_proptest_results}
\title{Convert prop.test to ARD}
\usage{
.format_proptest_results(by, variable, lst_tidy, ...)
}
\arguments{
\item{by}{(\code{string})\cr by column name}

\item{variable}{(\code{string})\cr variable column name}

\item{lst_tidy}{(named \code{list})\cr
list of tidied results constructed with \code{\link[cards:eval_capture_conditions]{eval_capture_conditions()}},
e.g. \code{eval_capture_conditions(t.test(mtcars$mpg ~ mtcars$am) |> broom::tidy())}.}

\item{...}{passed to \code{prop.test(...)}}
}
\value{
ARD data frame
}
\description{
Convert prop.test to ARD
}
\examples{
NULL

}
\keyword{internal}
