% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_surv_est.R
\name{.format_surv_est_results}
\alias{.format_surv_est_results}
\title{Convert t-test to ARD}
\usage{
.format_surv_est_results(tidy_stats, ...)
}
\arguments{
\item{...}{passed to \code{t.test(...)}}

\item{by}{(\code{string})\cr by column name}

\item{variable}{(\code{string})\cr variable column name}
}
\value{
ARD data frame
}
\description{
Convert t-test to ARD
}
\examples{
cardx:::.format_surv_est_results(
  broom::tidy(survfit(Surv(AVAL, CNSR) ~ TRTA, cards::ADTTE))
)
}
\keyword{internal}
