% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_wilcoxtest.R
\name{.format_wilcoxtest_results}
\alias{.format_wilcoxtest_results}
\title{Convert t-test to ARD}
\usage{
.format_wilcoxtest_results(by, variable, lst_tidy, paired, ...)
}
\arguments{
\item{by}{(\code{string})\cr by column name}

\item{variable}{(\code{string})\cr variable column name}

\item{lst_tidy}{(named \code{list})\cr
list of tidied results constructed with \code{eval_capture_conditions()},
e.g. \code{eval_capture_conditions(t.test(mtcars$mpg ~ mtcars$am) |> broom::tidy())}}

\item{paired}{a logical indicating whether you want a paired test.}

\item{...}{passed to \code{wilcox.test(...)}}
}
\value{
ARD data frame
}
\description{
Convert t-test to ARD
}
\examples{
cardx:::.format_wilcoxtest_results(
  by = "ARM",
  variable = "AGE",
  paired = FALSE,
  lst_tidy =
    cards::eval_capture_conditions(
      stats::wilcox.test(ADSL[["AGE"]] ~ ADSL[["ARM"]], paired = FALSE) |>
        broom::tidy()
    )
)
}
\keyword{internal}
