% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_ttest.R, R/ard_wilcoxtest.R
\name{.paired_data_pivot_wider}
\alias{.paired_data_pivot_wider}
\title{Convert long paired data to wide}
\usage{
.paired_data_pivot_wider(data, by, variable, id, env = rlang::caller_env())

.paired_data_pivot_wider(data, by, variable, id, env = rlang::caller_env())
}
\arguments{
\item{data}{(\code{data.frame})\cr a data frame that is one line per subject per group}

\item{by}{(\code{string})\cr by column name}

\item{variable}{(\code{string})\cr variable column name}

\item{id}{(\code{string})\cr subject id column name}

\item{env}{(\code{environment}) used for error messaging. Default is \code{rlang::caller_env()}}
}
\value{
a wide data frame

a wide data frame
}
\description{
Convert long paired data to wide

Convert long paired data to wide
}
\examples{
cards::ADSL[c("ARM", "AGE")] |>
  dplyr::filter(ARM \%in\% c("Placebo", "Xanomeline High Dose")) |>
  dplyr::mutate(.by = ARM, USUBJID = dplyr::row_number()) |>
  dplyr::arrange(USUBJID, ARM) |>
  cardx:::.paired_data_pivot_wider(by = "ARM", variable = "AGE", id = "USUBJID")
cards::ADSL[c("ARM", "AGE")] |>
  dplyr::filter(ARM \%in\% c("Placebo", "Xanomeline High Dose")) |>
  dplyr::mutate(.by = ARM, USUBJID = dplyr::row_number()) |>
  dplyr::arrange(USUBJID, ARM) |>
  cardx:::.paired_data_pivot_wider(by = "ARM", variable = "AGE", id = "USUBJID")
}
\keyword{internal}
