% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_survfit.R
\name{.process_survfit_probs}
\alias{.process_survfit_probs}
\title{Process Survival Fit For Quantile Estimates}
\usage{
.process_survfit_probs(x, probs)
}
\arguments{
\item{x}{(\code{\link[survival:survfit]{survival::survfit()}})\cr
a \code{\link[survival:survfit]{survival::survfit()}} object. See below for details.}

\item{probs}{(\code{numeric})\cr
a vector of probabilities with values in (0,1) specifying the survival quantiles to return.}
}
\value{
a tibble
}
\description{
Process Survival Fit For Quantile Estimates
}
\examples{
survfit(Surv(AVAL, CNSR) ~ TRTA, cards::ADTTE) |>
  cardx:::.process_survfit_probs(probs = c(0.25, 0.75))

}
\keyword{internal}
