% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_ci.R
\name{single_ci}
\alias{single_ci}
\alias{single_ci_mean}
\alias{single_ci_sd}
\alias{single_ci_var}
\title{Functions for Calculating Univariate Confidence Intervals}
\usage{
single_ci_mean(x, conf.level = 0.95, use_t = TRUE)

single_ci_sd(x, conf.level = 0.95)

single_ci_var(x, conf.level = 0.95)
}
\arguments{
\item{x}{vector of numeric values}

\item{conf.level}{(\code{numeric})\cr
a scalar in \verb{(0, 1)} indicating the confidence level.
Default is \code{0.95}}

\item{use_t}{(\code{logical})\cr use t-distribution vs normal distribution}
}
\value{
Confidence interval of a mean
}
\description{
Functions to calculate different univariate/single statistic confidence intervals for use in \code{ard_single_ci()}.
}
\section{Functions}{
\itemize{
\item \code{single_ci_mean()}: Calculates the mean confidenence interval by following
the usual textbook definition with either the normal or t distribution

\item \code{single_ci_sd()}: Calculates the standard deviation confidence interval by following
the usual textbook definition with the chi-sq distribution

\item \code{single_ci_var()}: Calculates the variance deviation confidence interval by following
the usual textbook definition with the chi-sq distribution

}}
\examples{
\dontshow{if (cards::is_pkg_installed("broom", reference_pkg = "cardx")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
x <- 1:10

single_ci_mean(x, conf.level = 0.9, use_t = TRUE)
single_ci_mean(x, conf.level = 0.9, use_t = FALSE)
single_ci_sd(x, conf.level = 0.9)
single_ci_var(x, conf.level = 0.9)
\dontshow{\}) # examplesIf}
}
