% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_aov.R
\name{ard_aov}
\alias{ard_aov}
\title{ARD ANOVA}
\usage{
ard_aov(data, by, variable)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame.}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name to compare by. Should be a factor variable.}

\item{variable}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name to be compared. Should be a continuous variable.}
}
\value{
ARD data frame
}
\description{
Analysis results data for Analysis of Variance.
Calculated with \code{aov(data[[variable]] ~ as.factor(data[[by]]), data = data, ...)}
}
\examples{
cards::ADSL |>
  ard_aov(by = "ARM", variable = "AGE")
}
