% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_anova.R
\name{ard_aov}
\alias{ard_aov}
\title{ARD ANOVA}
\usage{
ard_aov(data, by, variable, ...)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame.}

\item{by, variable}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column names to compare}

\item{...}{additional arguments passed to \code{aov(...)}}
}
\value{
ARD data frame
}
\description{
Analysis results data for Analysis of Variance.
Calculated with \code{aov(data[[variable]] ~ as.factor(data[[by]]), ...)}
}
\examples{
cards::ADSL |>
  ard_aov(by = "ARM", variable = "AGE")
}
