% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_aov.R
\name{ard_aov}
\alias{ard_aov}
\title{ARD ANOVA}
\usage{
ard_aov(x, ...)
}
\arguments{
\item{x}{regression model object}

\item{...}{arguments passed to \code{stats::aov(...)}}
}
\value{
ARD data frame
}
\description{
Analysis results data for Analysis of Variance.
Calculated with \code{stats::aov()}
}
\examples{
lm(AGE ~ ARM, data = cards::ADSL) |>
  ard_aov()
}
