% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_car_vif.R
\name{ard_car_vif}
\alias{ard_car_vif}
\title{Regression VIF ARD}
\usage{
ard_car_vif(x, ...)
}
\arguments{
\item{x}{regression model object
See car::vif() for details}

\item{...}{arguments passed to \code{car::vif(...)}}
}
\value{
data frame
}
\description{
Function takes a regression model object and returns the variance inflation factor (VIF)
using \code{\link[car:vif]{car::vif()}} and converts it to a ARD structure
}
\examples{
\dontshow{if (do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = "car"))) withAutoprint(\{ # examplesIf}
lm(AGE ~ ARM + SEX, data = cards::ADSL) |>
  ard_car_vif()
\dontshow{\}) # examplesIf}
}
