% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_incidence_rate.R
\name{ard_incidence_rate}
\alias{ard_incidence_rate}
\title{ARD Incidence Rate}
\usage{
ard_incidence_rate(
  data,
  time,
  count = NULL,
  id = NULL,
  by = NULL,
  strata = NULL,
  conf.level = 0.95,
  conf.type = c("normal", "normal-log", "exact", "byar"),
  units = c("years", "months", "weeks", "days"),
  n_person_years = 100
)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame.}

\item{time}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name of time at risk variable.}

\item{count}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name of variable indicating count of events that occurred. If \code{NULL}, each row in \code{data} is assumed to
correspond to a single event occurrence.}

\item{id}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name used to identify unique patients in \code{data}. If \code{NULL}, each row in \code{data} is assumed to correspond to
a unique patient.}

\item{by, strata}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
columns to tabulate by/stratify by for summary statistic
calculation. Arguments are similar, but with an important distinction:

\code{by}: results are calculated for \strong{all combinations} of the columns specified,
including unobserved combinations and unobserved factor levels.

\code{strata}: results are calculated for \strong{all \emph{observed} combinations} of the
columns specified.

Arguments may be used in conjunction with one another.}

\item{conf.level}{(\code{numeric})\cr
confidence level for the estimated incidence rate.}

\item{conf.type}{(\code{string})\cr
confidence interval type for the estimated incidence rate.

One of: \code{normal} (default), \code{normal-log}, \code{exact}, or \code{byar}.}

\item{units}{(\code{string})\cr
unit of time of values in \code{time}.

One of: \code{years} (default), \code{months}, \code{weeks}, or \code{days}}

\item{n_person_years}{(\code{numeric})\cr
number of person-years to estimate incidence rate for. Defaults to 100.}
}
\value{
an ARD data frame of class 'card'
}
\description{
Function takes a time at risk variable (\code{time}) and event count variable (\code{count}) and calculates the incidence
rate in person years.
}
\details{
Incidence rate is calculated as:

Total number of events that occurred / Total person-time at risk
}
\examples{
set.seed(1)
data <- data.frame(
  USUBJID = 1:100,
  TRTA = sample(LETTERS[1:3], 100, replace = TRUE),
  AETTE1 = abs(rnorm(100, mean = 0.5)),
  AETOT1 = sample(0:20, 100, replace = TRUE)
)

data |>
  ard_incidence_rate(time = AETTE1, count = AETOT1, id = USUBJID, by = TRTA)
}
