% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_onewaytest.R
\name{ard_onewaytest}
\alias{ard_onewaytest}
\title{ARD One-way Test}
\usage{
ard_onewaytest(x, ...)
}
\arguments{
\item{x}{regression model object}

\item{...}{additional arguments passed to \code{oneway.test(...)}}
}
\value{
ARD data frame
}
\description{
Analysis results data for Testing Equal Means in a One-Way Layout.
calculated with \code{oneway.test()}
}
\examples{
lm(AGE ~ ARM, data = cards::ADSL) |>
  ard_onewaytest()
}
