% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_onewaytest.R
\name{ard_onewaytest}
\alias{ard_onewaytest}
\title{ARD One-way Test}
\usage{
ard_onewaytest(data, by, variable, ...)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame.}

\item{by, variable}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column names to compare}

\item{...}{additional arguments passed to \code{oneway.test(...)}}
}
\value{
ARD data frame
}
\description{
Analysis results data for Testing Equal Means in a One-Way Layout.
Calculated with \code{oneway.test(data[[variable]] ~ as.factor(data[[by]]), ...)}
}
\examples{
cards::ADSL |>
  ard_onewaytest(by = "ARM", variable = "AGE")
}
