% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_proptest.R
\name{ard_proptest}
\alias{ard_proptest}
\title{ARD 2-sample proportion test}
\usage{
ard_proptest(data, by, variable, ...)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame.}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name to compare by}

\item{variable}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name to be compared. Must be a binary column coded as \code{TRUE}/\code{FALSE}
of \code{1}/\code{0}.}

\item{...}{arguments passed to \code{prop.test(...)}}
}
\value{
ARD data frame
}
\description{
Analysis results data for a 2-sample test or proportions using \code{\link[stats:prop.test]{stats::prop.test()}}.
}
\examples{
mtcars |>
  ard_proptest(by = vs, variable = am)
}
