% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_regression.R
\name{ard_regression_basic}
\alias{ard_regression_basic}
\title{Basic Regression ARD}
\usage{
ard_regression_basic(
  x,
  tidy_fun = broom.helpers::tidy_with_broom_or_parameters,
  ...
)
}
\arguments{
\item{x}{regression model object}

\item{tidy_fun}{(\code{function})\cr
a tidier. Default is \code{\link[broom.helpers:tidy_with_broom_or_parameters]{broom.helpers::tidy_with_broom_or_parameters}}}

\item{...}{Arguments passed to \code{broom.helpers::tidy_plus_plus()}}
}
\value{
data frame
}
\description{
A function that takes a regression model and provides basic statistics in an
ARD structure.
The default output is simpler than \code{ard_regression()}.
The function primarily matches regression terms to underlying variable names.
}
\examples{
lm(AGE ~ ARM, data = cards::ADSL) |>
  ard_regression_basic()

}
