% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_smd.R
\name{ard_smd}
\alias{ard_smd}
\title{ARD Standardized Mean Difference}
\usage{
ard_smd(data, by, variables, ...)
}
\arguments{
\item{data}{(\code{data.frame}/\code{survey.design})\cr
a data frame or object of class 'survey.design'
(typically created with \code{\link[survey:svydesign]{survey::svydesign()}}).}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name to compare by.}

\item{variables}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column names to be compared. Independent tests will be computed for
each variable.}

\item{...}{
  Arguments passed on to \code{\link[smd:smd]{smd::smd}}
  \describe{
    \item{\code{std.error}}{Logical indicator for computing standard errors using
\code{\link[smd]{compute_smd_var}}. Defaults to \code{FALSE}.}
    \item{\code{gref}}{an integer indicating which level of \code{g} to use as the reference
group. Defaults to \code{1}.}
  }}
}
\value{
ARD data frame
}
\description{
Standardized mean difference calculated via \code{\link[smd:smd]{smd::smd()}} with \code{na.rm = TRUE}.
}
\examples{
\dontshow{if (do.call(asNamespace("cards")$is_pkg_installed, list(pkg = "smd", reference_pkg = "cardx"))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ard_smd(cards::ADSL, by = ARM, variables = AGE, std.error = TRUE)
ard_smd(cards::ADSL, by = ARM, variables = AGEGR1, std.error = TRUE)
\dontshow{\}) # examplesIf}
}
