% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_total_n.survey.design.R
\name{ard_total_n.survey.design}
\alias{ard_total_n.survey.design}
\alias{ard_total_n.svyrep.design}
\title{ARD Total N}
\usage{
\method{ard_total_n}{survey.design}(data, ...)

\method{ard_total_n}{svyrep.design}(data, ...)
}
\arguments{
\item{data}{(\code{survey.design}) or (\code{svyrep.design})\cr
a survey design object or survey replicate object often created with
\code{\link[survey:svydesign]{survey::svydesign()}}, \code{\link[survey:as.svrepdesign]{survey::as.svrepdesign()}}, or
\code{\link[survey:svrepdesign]{survey::svrepdesign()}}}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
an ARD data frame of class 'card'
}
\description{
Returns the total N for a survey object.
The placeholder variable name returned in the object is \code{"..ard_total_n.."}
}
\examples{
\dontshow{if (cardx:::is_pkg_installed("survey")) withAutoprint(\{ # examplesIf}
svy_titanic <- survey::svydesign(~1, data = as.data.frame(Titanic), weights = ~Freq)

ard_total_n(svy_titanic)
\dontshow{\}) # examplesIf}
\dontshow{if (do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = "survey"))) withAutoprint(\{ # examplesIf}
# replicate-weight designs are also supported
data(api, package = "survey")
rclus1 <-
  survey::svydesign(id = ~dnum, weights = ~pw, data = apiclus1, fpc = ~fpc) |>
  survey::as.svrepdesign()

ard_total_n(rclus1)
\dontshow{\}) # examplesIf}
}
