% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_vif.R
\name{ard_vif}
\alias{ard_vif}
\title{Regression VIF ARD}
\usage{
ard_vif(x, statistic)
}
\arguments{
\item{x}{regression model object}

\item{statistic}{"VIF" (variance inflation factors, for models with no categorical terms)
or one of/combination of "GVIF" (generalized variance inflation factors),
"aGVIF" 'adjusted GVIF, i.e. GVIF^\link{1/(2*df)} and/or "df" (degrees of freedom).
See car::vif() for details}
}
\value{
data frame
}
\description{
Function takes a regression model object and converts it to a ARD
structure using the \code{broom.helpers} package.
}
\examples{
lm(AGE ~ ARM + SEX, data = cards::ADSL) |>
  ard_vif()
}
