% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_wilcoxtest.R
\name{ard_wilcoxtest}
\alias{ard_wilcoxtest}
\alias{ard_paired_wilcoxtest}
\title{ARD Wilcoxon Rank-Sum Test}
\usage{
ard_wilcoxtest(data, by, variables, ...)

ard_paired_wilcoxtest(data, by, variables, id, ...)
}
\arguments{
\item{data}{(\code{data.frame})\cr
a data frame. See below for details.}

\item{by}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
optional column name to compare by.}

\item{variables}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column names to be compared. Independent tests will be computed for
each variable.}

\item{...}{arguments passed to \code{wilcox.test(...)}}

\item{id}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
column name of the subject or participant ID.}
}
\value{
ARD data frame
}
\description{
Analysis results data for paired and non-paired Wilcoxon Rank-Sum tests.
}
\details{
For the \code{ard_wilcoxtest()} function, the data is expected to be one row per subject.
The data is passed as \code{wilcox.test(data[[variable]] ~ data[[by]], paired = FALSE, ...)}.

For the \code{ard_paired_wilcoxtest()} function, the data is expected to be one row
per subject per by level. Before the test is calculated, the data are
reshaped to a wide format to be one row per subject.
The data are then passed as
\verb{wilcox.test(x = data_wide[[<by level 1>]], y = data_wide[[<by level 2>]], paired = TRUE, ...)}.
}
\examples{
\dontshow{if (cards::is_pkg_installed("broom", reference_pkg = "cardx")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cards::ADSL |>
  dplyr::filter(ARM \%in\% c("Placebo", "Xanomeline High Dose")) |>
  ard_wilcoxtest(by = "ARM", variables = "AGE")

# constructing a paired data set,
# where patients receive both treatments
cards::ADSL[c("ARM", "AGE")] |>
  dplyr::filter(ARM \%in\% c("Placebo", "Xanomeline High Dose")) |>
  dplyr::mutate(.by = ARM, USUBJID = dplyr::row_number()) |>
  dplyr::arrange(USUBJID, ARM) |>
  ard_paired_wilcoxtest(by = ARM, variables = AGE, id = USUBJID)
\dontshow{\}) # examplesIf}
}
