% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ci.R
\name{boot_ci}
\alias{boot_ci}
\title{Function for Calculating Nonparametric Bootstrap Confidence Intervals}
\usage{
boot_ci(
  x,
  conf.level = 0.95,
  R = 2000,
  type = "all",
  stype = "i",
  statistic = function(x, i) mean(x[i], na.rm = TRUE)
)
}
\arguments{
\item{x}{(\code{numeric})\cr vector of numeric values, i.e. a numeric vector, or a logical vector which will be
interpreted as binary with values \code{c(0, 1)}.}

\item{conf.level}{(\code{numeric})\cr a scalar in \verb{(0, 1)} indicating the confidence level. Default is \code{0.95}.}

\item{R}{(\code{integer})\cr the number of bootstrap replicates. See the parameter description in \code{\link[boot:boot]{boot::boot()}}
for details.}

\item{type}{(\code{character})\cr a vector of character strings representing the type of intervals required. The value
should be any subset of the values \code{c("norm","basic", "stud", "perc", "bca")} or simply \code{"all"} which will
compute all five types of intervals.}

\item{stype}{(\code{string})\cr a character string indicating what the second argument of statistic represents.
Possible values of stype are \code{"i"} (indices - the default), \code{"f"} (frequencies), or \code{"w"} (weights).}

\item{statistic}{(\code{function})\cr a function which, when applied to data, returns a vector containing the
statistic(s) of interest. See the parameter description in \code{\link[boot:boot]{boot::boot()}} for details.}
}
\value{
Bootstrap confidence interval.
}
\description{
This function first applies \code{\link[boot:boot]{boot::boot()}} to convert the input data \code{x} to a \code{boot} object and then uses
\code{\link[boot:boot.ci]{boot::boot.ci()}} to calculate the corresponding nonparametric bootstrap confidence interval.
}
\examples{
\dontshow{if (do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = "boot", reference_pkg = "cardx"))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cards::ADSL |>
  cards::ard_continuous(
    variables = AGE,
    statistic = everything() ~ list(boot_ci = boot_ci)
  )
\dontshow{\}) # examplesIf}
}
