% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardx-package.R
\docType{package}
\name{cardx-package}
\alias{cardx}
\alias{cardx-package}
\title{cardx: Extra Analysis Results Data Utilities}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R Package to Supplement ARD Functions Found in {cards}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/cardx}
  \item Report bugs at \url{https://github.com/insightsengineering/cardx/issues}
}

}
\author{
\strong{Maintainer}: Daniel Sjoberg \email{sjobergd@gene.com}

Authors:
\itemize{
  \item Emily de la Rua \email{emily.de_la_rua@contractors.roche.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
