% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardx-package.R
\docType{package}
\name{cardx-package}
\alias{cardx}
\alias{cardx-package}
\title{cardx: Extra Analysis Results Data Utilities}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create extra Analysis Results Data (ARD) summary objects. The package supplements the simple ARD functions from the 'cards' package, exporting functions to put statistical results in the ARD format. These objects are used and re-used to construct summary tables, visualizations, and written reports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/cardx}
  \item \url{https://insightsengineering.github.io/cardx/}
  \item Report bugs at \url{https://github.com/insightsengineering/cardx/issues}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

Authors:
\itemize{
  \item Abinaya Yogasekaram \email{abinaya.yogasekaram@contractors.roche.com}
  \item Emily de la Rua \email{emily.de_la_rua@contractors.roche.com}
}

Other contributors:
\itemize{
  \item Malcolm Barrett (\href{https://orcid.org/0000-0003-0299-5825}{ORCID}) [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
