% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_aod_wald_test.R
\name{.extract_wald_results}
\alias{.extract_wald_results}
\title{Extract data from wald.test object}
\usage{
.extract_wald_results(wald_test)
}
\arguments{
\item{wald_test}{(\code{data.frame})\cr wald test object object from \code{aod::wald.test()}}
}
\value{
a data frame containing the wald test results.
}
\description{
Extract data from wald.test object
}
\keyword{internal}
