% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_survival_survfit.R
\name{.format_survfit_results}
\alias{.format_survfit_results}
\title{Convert Tidied Survival Fit to ARD}
\usage{
.format_survfit_results(tidy_survfit)
}
\value{
an ARD data frame of class 'card'
}
\description{
Convert Tidied Survival Fit to ARD
}
\examples{
\dontshow{if (do.call(asNamespace("cardx")$is_pkg_installed, list(pkg = c("survival", "broom")))) withAutoprint(\{ # examplesIf}
cardx:::.format_survfit_results(
  broom::tidy(survival::survfit(survival::Surv(AVAL, CNSR) ~ TRTA, cards::ADTTE))
)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
