% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_survfit.R
\name{.process_survfit_time}
\alias{.process_survfit_time}
\title{Process Survival Fit For Time Estimates}
\usage{
.process_survfit_time(x, times, type)
}
\arguments{
\item{x}{(\code{\link[survival:survfit]{survival::survfit()}})\cr
a \code{\link[survival:survfit]{survival::survfit()}} object. See below for details.}

\item{times}{(\code{numeric})\cr
a vector of times for which to return survival probabilities.}

\item{type}{(\code{character} or \code{NULL})\cr
type of statistic to report. Available for Kaplan-Meier time estimates only, otherwise \code{type}
is ignored. Default is \code{"survival"}.
Must be one of the following:\tabular{ll}{
   type \tab transformation \cr
   \code{"survival"} \tab \code{x} \cr
   \code{"risk"} \tab \code{1 - x} \cr
   \code{"cumhaz"} \tab \code{-log(x)} \cr
}}
}
\value{
a \code{tibble}
}
\description{
Process Survival Fit For Time Estimates
}
\examples{
survival::survfit(survival::Surv(AVAL, CNSR) ~ TRTA, cards::ADTTE) |>
  cardx:::.process_survfit_time(times = c(60, 180), type = "risk")

}
\keyword{internal}
