% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ard_survfit.R
\name{.process_survfit_time}
\alias{.process_survfit_time}
\title{Process Survival Fit For Time Estimates}
\usage{
.process_survfit_time(x, times, reverse)
}
\arguments{
\item{x}{(\code{\link[survival:survfit]{survival::survfit()}})\cr
a \code{\link[survival:survfit]{survival::survfit()}} object. See below for details.}

\item{times}{(\code{numeric})\cr
a vector of times for which to return survival probabilities.}

\item{reverse}{(\code{logical})\cr
Flip the probability reported, i.e. \code{1 - estimate}. Default is \code{FALSE}. Only applies when
\code{probs} is specified.}
}
\value{
a tibble
}
\description{
Process Survival Fit For Time Estimates
}
\examples{
survfit(Surv(AVAL, CNSR) ~ TRTA, cards::ADTTE) |>
  cardx:::.process_survfit_time(times = c(60, 180), reverse = FALSE)

}
\keyword{internal}
