% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_ci.R
\name{single_ci}
\alias{single_ci}
\title{Functions for Calculating Univariate Confidence Intervals}
\arguments{
\item{x}{vector of numeric values}
}
\value{
Confidence interval of a mean
}
\description{
Functions to calculate different univariate/single statistic confidence intervals for use in \code{ard_single_ci()}.
}
\examples{
\dontshow{if (cards::is_pkg_installed("broom", reference_pkg = "cardx")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
x <- 1:10

single_ci_mean(x, conf.level = 0.9, use_t = TRUE)
single_ci_mean(x, conf.level = 0.9, use_t = FALSE)
\dontshow{\}) # examplesIf}
}
